+++ config/dev.exs
@@ -2,6 +2,6 @@ use Mix.Config
 
 config :ex_cldr,
   default_locale: "en",
-  locales: ["root", "fr", "en", "bs", "pl", "ru", "th", "he"],
+  locales: ["root", "fr", "en", "bs", "pl", "ru", "th", "he", "af", "af-NA"],
   gettext: Cldr.Gettext,
   precompile_number_formats: ["¤¤#,##0.##"]